Implementing library links in c++.

+++ src/maxflow/graph.h
@@ -66,9 +66,9 @@ public:
 
 	/* Type of edge weights.
 	   Can be changed to char, int, float, double, ... */
-	typedef float captype;
+	typedef int captype;
 	/* Type of total flow */
-	typedef double flowtype;
+	typedef long flowtype;
 
 	typedef void * node_id;
